feat: use sqlc 1.23.0

+++ internal/controller/controller_helper/item_helper.go
@@ -211,13 +211,13 @@ func generateQueryItemsCondition(uid pgtype.UUID, b model.GetItemsRequestBody) (
 		condition.Limit = 10
 		b.Size = 10
 	} else {
-		condition.Limit = int32(b.Size)
+		condition.Limit = b.Size
 	}
 	if b.Page == 0 {
 		condition.Offset = 0
 		b.Page = 1
 	} else {
-		condition.Offset = int32((b.Page - 1) * b.Size)
+		condition.Offset = (b.Page - 1) * b.Size
 	}
 
 	after, before := generateDateRange(b.HappenedAfter, b.HappenedBefore)
